refactor(transcription): share document text/vision/legacy pipeline
